working implementation of always getting and using latest CLI
