+++ pkgs/npm/src/cli.ts
@@ -1,14 +1,197 @@
 #!/usr/bin/env node
-
-import * as path from "path";
+import AdmZip from "adm-zip";
+import axios from "axios";
 import * as child_process from "child_process";
+import { promises as fsPromise } from "fs";
+import * as os from "os";
+import * as path from "path";
+import * as tar from "tar";
 import { promisify } from "util";
-import { downloadAppArchive } from "./install";
 
-const exec = promisify(child_process.exec);
 const EXECUTABLE = "defang";
 
-function getPathToExecutable(): string {
+const exec = promisify(child_process.exec);
+
+async function getLatestVersion(): Promise<string> {
+  const url = "https://api.github.com/repos/DefangLabs/defang/releases/latest";
+  const response = await axios.get(url);
+  if (response.status !== 200) {
+    throw new Error(
+      `Failed to get latest version from GitHub. Status code: ${response.status}`
+    );
+  }
+
+  return response.data.tag_name.replace("v", "").trim();
+}
+
+async function downloadAppArchive(
+  version: string,
+  archiveFilename: string,
+  outputPath: string
+): Promise<string> {
+  const repo = "DefangLabs/defang";
+  const downloadUrl = `https://github.com/${repo}/releases/download/v${version}/${archiveFilename}`;
+  const downloadTargetFile = path.join(outputPath, archiveFilename);
+
+  return downloadFile(downloadUrl, downloadTargetFile);
+}
+
+async function downloadFile(
+  downloadUrl: string,
+  downloadTargetFile: string
+): Promise<string> {
+  try {
+    console.log(`Downloading ${downloadUrl}`);
+    const response = await axios.get(downloadUrl, {
+      responseType: "arraybuffer",
+      headers: {
+        "Content-Type": "application/octet-stream",
+      },
+    });
+
+    if (response?.data === undefined) {
+      throw new Error(
+        `Failed to download ${downloadUrl}. No data in response.`
+      );
+    }
+
+    await fsPromise.writeFile(downloadTargetFile, response.data);
+
+    return downloadTargetFile;
+  } catch (error) {
+    console.error(error);
+    await fsPromise.unlink(downloadTargetFile);
+    return "";
+  }
+}
+
+function extractArchive(archiveFilePath: string, outputPath: string): boolean {
+  let result = false;
+
+  console.log(`Extracting ${archiveFilePath}`);
+  const ext = path.extname(archiveFilePath).toLocaleLowerCase();
+  switch (ext) {
+    case ".zip":
+      result = extractZip(archiveFilePath, outputPath);
+      break;
+    case ".gz":
+      result = extractTarGz(archiveFilePath, outputPath);
+      break;
+    default:
+      throw new Error(`Unsupported archive extension: ${ext}`);
+  }
+
+  return result;
+}
+
+function extractZip(zipPath: string, outputPath: string): boolean {
+  try {
+    const zip = new AdmZip(zipPath);
+    zip.extractAllTo(outputPath, true, true);
+    return true;
+  } catch (error) {
+    console.error(`An error occurred during zip extraction: ${error}`);
+    return false;
+  }
+}
+
+function extractTarGz(tarGzFilePath: string, outputPath: string): boolean {
+  try {
+    tar.extract(
+      {
+        cwd: outputPath,
+        file: tarGzFilePath,
+        sync: true,
+        strict: true,
+      },
+      [EXECUTABLE]
+    );
+    return true;
+  } catch (error) {
+    console.error(`An error occurred during tar.gz extraction: ${error}`);
+    return false;
+  }
+}
+
+async function deleteArchive(archiveFilePath: string): Promise<void> {
+  await fsPromise.unlink(archiveFilePath);
+}
+
+async function getVersion(filename: string): Promise<string> {
+  const data = await fsPromise.readFile(filename, "utf8");
+  const pkg = JSON.parse(data);
+  return pkg.version;
+}
+
+function getAppArchiveFilename(
+  version: string,
+  platform: string,
+  arch: string
+): string {
+  let compression = "zip";
+  switch (platform) {
+    case "windows":
+      platform = "windows";
+      break;
+    case "linux":
+      platform = "linux";
+      compression = "tar.gz";
+      break;
+    case "darwin":
+      platform = "macOS";
+      break;
+    default:
+      throw new Error(`Unsupported operating system: ${platform}`);
+  }
+
+  switch (arch) {
+    case "x64":
+      arch = "amd64";
+      break;
+    case "arm64":
+      arch = "arm64";
+      break;
+    default:
+      throw new Error(`Unsupported architecture: ${arch}`);
+  }
+
+  if (platform === "macOS") {
+    return `defang_${version}_${platform}.${compression}`;
+  }
+  return `defang_${version}_${platform}_${arch}.${compression}`;
+}
+
+async function install(version: string, saveDirectory: string) {
+  try {
+    console.log(`Starting install of defang cli`);
+
+    const filename = getAppArchiveFilename(version, os.platform(), os.arch());
+    const archiveFile = await downloadAppArchive(
+      version,
+      filename,
+      saveDirectory
+    );
+
+    if (archiveFile.length === 0) {
+      throw new Error(`Failed to download ${filename}`);
+    }
+
+    // Because the releases are compressed tar.gz or .zip we need to
+    // uncompress them to the ./bin directory in the package in node_modules.
+    const result = extractArchive(archiveFile, saveDirectory);
+    if (result === false) {
+      throw new Error(`Failed to install binaries!`);
+    }
+    console.log(`Successfully installed defang cli!`);
+    // Delete the downloaded archive since we have successfully downloaded
+    // and uncompressed it.
+    await deleteArchive(archiveFile);
+  } catch (error) {
+    console.error(error);
+  }
+}
+
+function getPathToExecutable(): string | null {
   let extension = "";
   if (["win32", "cygwin"].includes(process.platform)) {
     extension = ".exe";
@@ -18,39 +201,55 @@ function getPathToExecutable(): string {
   try {
     return require.resolve(executablePath);
   } catch (e) {
-    throw new Error(`Could not find application binary at ${executablePath}.`);
+    return null;
   }
 }
 
-function extractCLIVersions(output: string): {
+function extractCLIVersions(versionInfo: string): {
   defangCLI: string;
   latestCLI: string;
 } {
-  const defangCLIRegex = /Defang CLI:\s+(\w+)/;
-  const latestCLIRegex = /Latest CLI:\s+(\S+)/;
+  const versionRegex = /\d+\.\d+\.\d+/g;
+  const matches = versionInfo.match(versionRegex);
 
-  const defangCLIMatch = output.match(defangCLIRegex);
-  const latestCLIMatch = output.match(latestCLIRegex);
-
-  if (defangCLIMatch && latestCLIMatch) {
+  if (matches != null && matches.length >= 2) {
     return {
-      defangCLI: defangCLIMatch[1],
-      latestCLI: latestCLIMatch[1],
+      defangCLI: matches[0],
+      latestCLI: matches[0],
     };
   } else {
     throw new Error("Could not extract CLI versions from the output.");
   }
 }
 
-async function getVersionInfo(): Promise<{ current: string; latest: string }> {
-  const execPath = getPathToExecutable();
+type VersionInfo = {
+  current: string | null;
+  latest: string | null;
+};
 
-  // Exec output contains both stderr and stdout outputs
-  const versionInfo = await exec(execPath + " --version");
+async function getVersionInfo(): Promise<VersionInfo> {
+  let result: VersionInfo = { current: null, latest: null };
+  try {
+    const execPath = getPathToExecutable();
 
-  const result = extractCLIVersions(versionInfo.stdout);
+    if (!execPath) {
+      // there is no executable, so we can't get the version info from the CLI
+      const latestVersion = await getLatestVersion();
 
-  return { current: result.defangCLI, latest: result.latestCLI };
+      return { current: null, latest: latestVersion };
+    }
+
+    // Exec output contains both stderr and stdout outputs
+    const versionInfo = await exec(execPath + " version");
+
+    const verInfo = extractCLIVersions(versionInfo.stdout);
+    result.current = verInfo.defangCLI;
+    result.latest = verInfo.latestCLI;
+  } catch (error) {
+    console.error(error);
+  }
+
+  return result;
 }
 
 // js wrapper to use by npx or npm exec, this will call the defang binary with
@@ -59,18 +258,25 @@ async function getVersionInfo(): Promise<{ current: string; latest: string }> {
 // the package name i.e. defang.
 async function run(): Promise<void> {
   try {
-    const { current, latest } = await getVersionInfo();
+    const { current, latest }: VersionInfo = await getVersionInfo();
 
-    if (current != latest) {
-      // download and install the latest version of defang cli
-      downloadAppArchive();
+    // get the latest version of defang cli if not already installed
+    if (latest != null && (current == null || current != latest)) {
+      await install(latest, __dirname);
     }
 
+    // execute the defang binary with the arguments passed to the npx line.
     const args = process.argv.slice(2);
-    const processResult = child_process.spawnSync(getPathToExecutable(), args, {
+    const pathToExec = getPathToExecutable();
+    if (!pathToExec) {
+      throw new Error("Could not find the defang executable.");
+    }
+
+    const processResult = child_process.spawnSync(pathToExec, args, {
       stdio: "inherit",
     });
 
+    // if there was an error, print it to the console.
     processResult.error && console.error(processResult.error);
     process.exitCode = processResult.status ?? 1;
   } catch (error) {
